feat(misc): move run-commands executor to nx
