+++ packages/nx/src/executors/run-commands/schema.json
@@ -7,15 +7,23 @@
   "presets": [
     {
       "name": "Arguments forwarding",
-      "keys": ["commands"]
+      "keys": [
+        "commands"
+      ]
     },
     {
       "name": "Custom done conditions",
-      "keys": ["commands", "readyWhen"]
+      "keys": [
+        "commands",
+        "readyWhen"
+      ]
     },
     {
       "name": "Setting the cwd",
-      "keys": ["commands", "cwd"]
+      "keys": [
+        "commands",
+        "cwd"
+      ]
     }
   ],
   "properties": {
@@ -73,7 +81,9 @@
               }
             },
             "additionalProperties": false,
-            "required": ["command"]
+            "required": [
+              "command"
+            ]
           },
           {
             "type": "string"
